Добавил accrual worker pool для рассчета вознаграждения за заказы

+++ internal/accrual/handler/api_orders.go
@@ -16,8 +16,8 @@ import (
 )
 
 type reqGood struct {
-	Description string `json:"description"`
-	Price       int    `json:"price"`
+	Description string  `json:"description"`
+	Price       float64 `json:"price"`
 }
 
 type reqOrder struct {
